Task #3168: Use the bit mode from the first RSP board, but check all of them for consistency

+++ branches/MAC-8bitmode-Task3168/MAC/APL/StationCU/src/ClockControl/ClockControl.h
@@ -127,7 +127,7 @@ private:
 	int32					itsClock;
 
 	void*					itsBitmodeSubscription;
-	int32					itsBitmode;
+	uint16					itsBitmode;
 
 	bool					itsSplitterRequest;
 	bitset<MAX_RSPBOARDS>	itsSplitters;
